requirements.txt
h5py==3.8.0
imgaug==0.4.0
ipython==8.12.0
keras==2.12.0
matplotlib==3.7.1
numpy==1.22.4
Pillow==9.5.0
pip==23.0.1
pycocotools==2.0.6
scikit_image==0.19.3
setuptools==67.6.1
skimage==0.0
tensorflow==2.12.0
torch==2.0.0+cu118
torchvision==0.15.1+cu118
cython
protobuf